+++ packages/@expo/cli/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Enable async require by default. ([#36405](https://github.com/expo/expo/pull/36405) by [@EvanBacon](https://github.com/EvanBacon))
 - Support JSON output for install check. ([#37318](https://github.com/expo/expo/pull/37318) by [@betomoedano](https://github.com/betomoedano))
+- Add support for `--` separator in `expo run <platform>`, `expo run:android` and `expo run:ios` commands, to allow running commands like `expo run android -d -- /path/to/my_app` ([#37301](https://github.com/expo/expo/pull/37301) by [@makl11](https://github.com/makl11))
 
 ### 🐛 Bug fixes